Добавляет модульность и будущую файловую структуру
